requirements.txt
# Core libraries
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Financial data
yfinance>=0.1.70
alpaca-trade-api>=2.3.0
ccxt>=1.60.0

# Statistical and machine learning
scipy>=1.7.0
statsmodels>=0.13.0
scikit-learn>=1.0.0
arch>=5.0.0

# Visualization
plotly>=5.0.0

# Database
sqlalchemy>=1.4.0
pymongo>=4.0.0
influxdb3-python>=0.10.0
certifi>=2023.7.22
requests>=2.27.0

# Network analysis
networkx>=2.6.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0

# Additional libraries
python-binance==1.0.16
fastapi==0.78.0
uvicorn==0.17.6
pytest==7.1.2
dash==2.5.0
dash-bootstrap-components==1.1.0 